examples/travel-agent/requirements.txt
google-generativeai==0.3.2
tavily-python==0.3.0
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.6.0
